Make Logoot sort itself on `insert` and `remove`
This is layering the ground to overall speed improvements, much needed.

Sorted is probably going to become an Array, to provide better
`insertAt` support and performance right in the data structure.

This is to avoid implementing a proper sorted hash map (TreeMap),
which would probably have better performance but not something I
want to deal just now.